Create Scrape_bills script
This file contains code to create a dataframe called 'clean_dataframe' with data about all the bills introduced to the Ecuadorian national Assembly in the 2009-2014 period. Each observation is a bill across all the possible legislative stages and contains a link to download the corresponding pdf document for each available stage. This script also contains download_function, which download the PDFs links from the dataframe.